Safari-609.1.11.1/LayoutTests/http/tests/workers/service/self_registration_update-expected.txt

* Add basic testing for ServiceWorkerGlobalScope.registration.update()

PASS: registration was successfuly updated by the service worker
PASS: service worker became redundant
PASS: Old worker is still the registration's active worker
PASS: New worker should be the registration's waiting worker
